// frontend/src/models/Fish.ts
export interface Fish {
  name: string
  donated: boolean
  icon_image: string
  critterpedia_image: string
  furniture_image: string
  sell: number
  where_how: string
  shadow_size: string
  catch_difficulty: string
  vision: string
  total_catches_to_unlock: number
  spawn_rates: string
  nh_jan: string
  nh_feb: string
  nh_mar: string
  nh_apr: string
  nh_may: string
  nh_jun: string
  nh_jul: string
  nh_aug: string
  nh_sep: string
  nh_oct: string
  nh_nov: string
  nh_dec: string
  sh_jan: string
  sh_feb: string
  sh_mar: string
  sh_apr: string
  sh_may: string
  sh_jun: string
  sh_jul: string
  sh_aug: string
  sh_sep: string
  sh_oct: string
  sh_nov: string
  sh_dec: string
  size: string
  surface: string
  description: string
  catch_phrase: string
  hha_base_points: number
  hha_category: string
  color1: string
  color2: string
  lighting_type: string
  id: string
  critterpedia_order: number
  cj_sell_price: number
  rarity: string
}